// adapters/implementations/codex-app-server/src/protocol/generated/v2/ThreadTokenUsageUpdatedNotification.ts
// GENERATED CODE! DO NOT MODIFY BY HAND!

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ThreadTokenUsage } from './ThreadTokenUsage';

export type ThreadTokenUsageUpdatedNotification = { threadId: string; turnId: string; tokenUsage: ThreadTokenUsage };